Zoe/requirements.txt
discord.py==1.7.3
dpytest==0.5.3
pytest==6.2.4
python-dotenv==0.18.0